# backend/models.py
"""
models.py — All Pydantic request/response models for every endpoint.

Pydantic enforces that Android sends us exactly the right shape of data.
If Android sends a malformed request, Pydantic rejects it automatically
with a clear error message — the bad data NEVER reaches our AI logic.
"""
from pydantic import BaseModel, Field
from typing import List, Optional, Any, Dict
from enum import Enum


# ═══════════════════════════════════════════════════════════════════════════
# SHARED ENUMS — used across multiple endpoints
# ═══════════════════════════════════════════════════════════════════════════

class ReplayDecision(str, Enum):
    EXECUTE_PLAN = "EXECUTE_PLAN"
    ASK_USER = "ASK_USER"
    UNKNOWN_INTENT = "UNKNOWN_INTENT"
    STOP = "STOP"

class RecoveryDecision(str, Enum):
    CONTINUE = "CONTINUE"
    RETRY_WITH_ALTERNATIVE = "RETRY_WITH_ALTERNATIVE"
    ASK_USER = "ASK_USER"
    STOP = "STOP"

class Intent(str, Enum):
    ORDER_FOOD = "order_food"
    SEARCH_PRODUCT = "search_product"
    ADD_TO_CART = "add_to_cart"
    UNKNOWN_INTENT = "unknown_intent"


# ═══════════════════════════════════════════════════════════════════════════
# SHARED SUB-MODELS
# ═══════════════════════════════════════════════════════════════════════════

class ActionStep(BaseModel):
    """One semantic action observed by Android during teaching or replay."""
    action: str = Field(..., description="e.g. SEARCH, CLICK, SET_QUANTITY, SCROLL, BACK")
    targetText: Optional[str] = Field(None, description="The visible text of the UI element")
    targetContentDesc: Optional[str] = Field(None, description="The content description of the element")
    targetResourceId: Optional[str] = Field(None, description="The Android resource ID of the element")
    value: Optional[str] = Field(None, description="The value typed or selected")
    success: Optional[bool] = Field(None, description="Whether this action succeeded on the device")

class UITargetMetadata(BaseModel):
    """Semantic description of a UI element observed by Android's AccessibilityService."""
    resourceId: Optional[str] = None
    text: Optional[str] = None
    contentDescription: Optional[str] = None
    viewClass: Optional[str] = None
    isClickable: Optional[bool] = None
    screenContext: Optional[str] = None  # e.g. "search_results", "cart_page"

class SlotDefinition(BaseModel):
    """Defines one variable slot in a learned flow."""
    name: str
    slotType: str = Field(..., description="e.g. restaurant, item, quantity, address, query")
    required: bool = True
    defaultValue: Optional[str] = None
    exampleValues: Optional[List[str]] = None

class GeneralizedStep(BaseModel):
    """One step in a generalized, reusable flow. Values may contain {{slot}} placeholders."""
    action: str
    targetText: Optional[str] = None
    value: Optional[str] = None
    targetMetadata: Optional[UITargetMetadata] = None
    isInvariant: bool = Field(False, description="True if this step never changes across replays")
    stopBefore: bool = Field(False, description="True if Android must STOP before executing this")


# ═══════════════════════════════════════════════════════════════════════════
# POST /v1/teach
# ═══════════════════════════════════════════════════════════════════════════

class TeachRequest(BaseModel):
    mode: str = Field(default="TEACH")
    requestId: str = Field(..., description="Unique ID for this request, generated by Android")
    sessionId: Optional[str] = None
    appPackage: str = Field(..., description="e.g. com.zomato.android")
    appName: Optional[str] = Field(None, description="e.g. Zomato")
    utterance: str = Field(..., description="Exact spoken command from the user")
    demonstration: List[ActionStep] = Field(..., min_length=1)
    teachingMetadata: Optional[Dict[str, Any]] = None
    schemaVersion: str = Field(default="v1")

class TeachResponse(BaseModel):
    status: str
    flowId: str
    intent: str
    slots: Dict[str, Any]
    slotDefinitions: List[SlotDefinition]
    generalizedSteps: List[GeneralizedStep]
    confidence: float = Field(ge=0.0, le=1.0)
    clarificationNeeded: bool = False
    clarificationQuestion: Optional[str] = None
    schemaVersion: str = "v1"
    debugSummary: Optional[str] = None  # Human-readable explanation for debugging


# ═══════════════════════════════════════════════════════════════════════════
# POST /v1/replay/plan
# ═══════════════════════════════════════════════════════════════════════════

class ReplayPlanRequest(BaseModel):
    requestId: str
    utterance: str = Field(..., description="The new command to match to a saved flow")
    appPackage: Optional[str] = None
    currentUISummary: Optional[str] = Field(None, description="Compact description of current screen")
    sessionContext: Optional[Dict[str, Any]] = None
    schemaVersion: str = "v1"

class ReplayPlanResponse(BaseModel):
    decision: ReplayDecision
    flowId: Optional[str] = None
    resolvedSlots: Optional[Dict[str, Any]] = None
    actions: Optional[List[GeneralizedStep]] = None
    clarificationQuestion: Optional[str] = None
    message: str
    confidence: Optional[float] = None
    schemaVersion: str = "v1"


# ═══════════════════════════════════════════════════════════════════════════
# POST /v1/recovery
# ═══════════════════════════════════════════════════════════════════════════

class RecoveryRequest(BaseModel):
    requestId: str
    flowId: str
    currentStepIndex: int
    expectedAction: ActionStep
    currentUISummary: str = Field(..., description="Compact description of what Android sees now")
    lastActionResult: str = Field(..., description="What happened when the last action was tried")
    failureReason: Optional[str] = None
    schemaVersion: str = "v1"

class RecoveryResponse(BaseModel):
    decision: RecoveryDecision
    suggestedAction: Optional[GeneralizedStep] = None
    reason: str
    confidence: float = Field(ge=0.0, le=1.0)
    nextExpectedState: Optional[str] = None
    clarificationQuestion: Optional[str] = None
    schemaVersion: str = "v1"


# ═══════════════════════════════════════════════════════════════════════════
# POST /v1/step-result
# ═══════════════════════════════════════════════════════════════════════════

class StepResultRequest(BaseModel):
    requestId: str
    flowId: str
    stepId: int
    action: ActionStep
    success: bool
    observedState: Optional[str] = None
    failureReason: Optional[str] = None
    schemaVersion: str = "v1"

class StepResultResponse(BaseModel):
    acknowledged: bool
    message: str
    schemaVersion: str = "v1"


# ═══════════════════════════════════════════════════════════════════════════
# POST /v1/flows/match
# ═══════════════════════════════════════════════════════════════════════════

class FlowMatchRequest(BaseModel):
    requestId: str
    utterance: str
    appPackage: Optional[str] = None
    schemaVersion: str = "v1"

class FlowMatchResponse(BaseModel):
    matched: bool
    flowId: Optional[str] = None
    extractedSlots: Optional[Dict[str, Any]] = None
    confidence: Optional[float] = None
    ambiguous: bool = False
    clarificationQuestion: Optional[str] = None
    message: str
    schemaVersion: str = "v1"


# ═══════════════════════════════════════════════════════════════════════════
# GET /health
# ═══════════════════════════════════════════════════════════════════════════

class HealthResponse(BaseModel):
    status: str
    version: str
    schemaVersion: str
    flowsStored: int